Split up app in multiple files - move endpoints to routes.js - rename index.js to app.js
